cleaned up a bit
git-svn-id: https://scm.mmbase.org/mmbase/trunk@45316 91430c5f-d668-0410-9e7d-dd663366dcfd

+++ bridge/src/main/java/org/mmbase/datatypes/DoubleDataType.java
@@ -26,8 +26,8 @@ public class DoubleDataType extends NumberDataType<Double> {
      */
     public DoubleDataType(String name, boolean primitive) {
         super(name, primitive ? Double.TYPE : Double.class);
-        setMin(Double.valueOf(-1 * Double.MAX_VALUE), true);
-        setMax(Double.valueOf(Double.MAX_VALUE), true);
+        setMin(-1 * Double.MAX_VALUE, true);
+        setMax(Double.MAX_VALUE, true);
     }
 
     /**
